Scrapped velocity in favor of rigid body component
I figure it would be more tedious to individually add a velocity and acceleration and mass and etc. components to each new entity each time when it is a given (for my purposes at least) that they will have all of these components, so I decided to lump them into each other under a RigidBody component.