Modifications to the player action system
Player action system will likely become the player state system, still working out the details...

Two different entities in test sim now, though this makes noticeable that the input system does not work for multiple players, only one player can move at a time. This might mean that the InputSystem will become a component and merge with the player input listener... Though it would be a hefty component and be difficult to manage.