Scene cleanup and creation of game class
The scene has been heavily trimmed, it now only holds on to the components and entities in the scene and does nothing with them.

The systems belong to a game class now and the instance of the game is updated each frame instead of the scene.

No more heap allocations using new.