Added Y sorting for sprites
sprites with a higher Y coordinate (lower down) will be rendered last, causing them to appear on top of other sprites. Layers are now a field in the sprite component allowing for way more than 16 layers, and scene has reverted to how it used to look mostly. More work to be done in cleaning up and organizing this new way of handling rendering.