Added a follow camera
Camera now follows the player. Camera following is handled by a CameraManager component and FollowCamera component which will be assigned to the camera entity in the scene, each scene will have 1 camera entity which will have a reference to the camera stored in the Game. I chose to do it this way to manage how many follow camera components exist, this way, you cannot add more than one as there is only one camera and you cannot get a reference to the camera the game uses.