Created tag component
The directionality, entity type, entity group, state, and other information will be contained in the first "Mandatory Component", the Tag. The tag component will be assigned to each entity upon instantiation, thus making it so ALL entities must have this component, it will hold all of the tags that describe the entity.

This will make conditional logic much easier as the tags can be used to decide what needs to happen.

I chose to not make many different component types for each tag field as doing so would quickly lead to most components being ignored by systems and some systems will require very specific tags which introduces more complexity. Doing it this way sort of generalizes the tag listing and allows systems to assume the tags they work on will always be present.