Reworked velocity
The units provided for all distance will simply be in meters, where each pixel is 1cm and so 100 pixels is 1 meter. So, the displacement provided by the velocity is now scaled up by PIXEL_SIZE * 100 before being applied. This makes the friction act much more like expected and removes the need for the stopping force.