Removed stopping force component
As well the movement controller system has been simplified since the logic is now the same for both movement types. Sprites now also take a xScale and yScale instead of an exact size in pixels to render at, this is then multiplied by PIXEL_SIZE and and the source rects width and height respectively for the destination rect.