Impulse tweak
Impulse is now calculated properly by dividing the force by the time and only applying it for one frame then removing itself.

The issue with this is that the speed limiter will limit the effect of the impulse so I will need a better way to separate the internal forces for movement and such from the external forces.