Minor changes to impulse and added resizable windows
Window is now fully resizable and game mechanics no longer depend directly on the pixel size.

Impulse and the movement controller are still not where I want them to be. Impulse works but it isn't very snappy and is finicky to get exactly what I want. Likewise, the movement controller will still stop applying force if the velocity hits a threshold, regardless of if that was because of a movement caused by the player or not.

Intuitively, if a force hits you in the direction of movement, you should go much faster in that same direction, regardless of how fast you can actually run by yourself. This is not the case here.